pyproject.toml
[project]
name = "light-control"
version = "0.0.1"
description = "DX-Light imitation"
readme = "README.md"
authors = [
    { name = "tan", email = "tanmaster95@hotmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "hidapi>=0.15.0",
    "pyside6",
]


[project.scripts]
light-control = "light_control:main"


[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"


[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff",
]

